Move Arity to its own namespace
