+++ src/Language/Parser/Symbol/Assignment.php
@@ -11,7 +11,7 @@
 
 namespace Symbiont\Language\Parser\Symbol;
 
-use Symbiont\Language\Ast\Node\Arity;
+use Symbiont\Language\Ast\Node\Arity\Arity;
 use Symbiont\Language\Ast\Node\AssignmentNode;
 use Symbiont\Language\Ast\Node\LiteralNodeInterface;
 use Symbiont\Language\Ast\Node\NodeInterface;
